+++ tests/screenshot/apps/components/TooltipDecorator.js
@@ -6,9 +6,26 @@ import {withConfig} from './utils';
 const TooltipButton = TooltipDecorator({tooltipDestinationProp: 'decoration'}, Button);
 
 const TooltipDecoratorTests = [
-	<TooltipButton tooltipText="tooltip">Click me</TooltipButton>,
-	<TooltipButton tooltipDelay={100} tooltipText="tooltip">Click me</TooltipButton>,
-	<TooltipButton tooltipPosition="left bottom" tooltipText="tooltip">Click me</TooltipButton>,
+	{
+		component: <TooltipButton tooltipText="tooltip">Click me</TooltipButton>,
+		focus: true
+	},
+	{
+		component: <TooltipButton tooltipDelay={100} tooltipText="tooltip">Click me</TooltipButton>,
+		focus: true
+	},
+	{
+		component: <TooltipButton tooltipPosition="left bottom" tooltipText="tooltip">Click me</TooltipButton>,
+		focus: true
+	},
+	{
+		component: <TooltipButton tooltipText="Tooltip to test tooltipWidth prop" tooltipWidth={50}>Click me</TooltipButton>,
+		focus: true
+	},
+	{
+		component: <TooltipButton tooltipText="Long tooltip to test that Marquee applies" tooltipMarquee tooltipWidth={150}>Click me</TooltipButton>,
+		focus: true
+	},
 
 	// RTL
 	...withConfig({